+++ www/modules/bookmarks/language/HU.php
@@ -1,0 +1,18 @@
+<?php
+return array (
+  'noEmployeesToDisplay' => 'noEmployeesToDisplay',
+  'Category' => 'Kategória',
+  'Group bookmarks in main start menu (Browser reload required)' => 'Csoport könyvjelzõlistáját fő Start menüben (Böngésző frissítése szükséges)',
+  'Administrate categories' => 'adminisztrációja kategória',
+  'Website description.' => 'Weboldal leírása.',
+  'items' => 'példány',
+  'item' => 'tétel',
+  'Toggle view' => 'toggle nézet',
+  'Are you sure you want to delete this bookmark?' => 'Biztos, hogy törölni szeretné ezt a könyvjelzőt?',
+  'Bookmarks' => 'Könyvjelzők',
+  'Choose icon for bookmark' => 'Válasszon ikont könyvjelző',
+  'Upload Logo' => 'Logó feltöltése',
+  'Bookmark' => 'Könyvjelző',
+  'Open in new browser tab' => 'Megnyitás új böngésző lapon',
+  'Behave as a module (Browser reload required)' => 'Viselkedjen úgy, mint egy modul (Böngésző frissítése szükséges)',
+);