+++ www/modules/comments/language/id.php
@@ -1,0 +1,9 @@
+<?php
+return array (
+  'Category' => 'Kategori',
+  'Manage categories' => 'Kelola kategori',
+  'Comments' => 'Komentar',
+  '{author} wrote at {date}' => '{author} menulis pada {date}',
+  'Action date' => 'Tanggal tindakan',
+  'Browse comments' => 'Telusuri komentar',
+);