+++ www/modules/customcss/language/HU.php
@@ -1,0 +1,4 @@
+<?php
+return array (
+  'Custom CSS styles and scripts' => 'Egyéni CSS stílusok és scriptek',
+);