+++ www/modules/freebusypermissions/language/id.php
@@ -1,0 +1,5 @@
+<?php
+return array (
+  'Accessrights for free busy info' => 'Accessrights untuk info sibuk bebas',
+  'Grant access to users and and groups to allow access to your free busy information. These users will also be able to schedule tentative events directly into your calendar.' => 'Berikan akses kepada pengguna dan grup untuk memungkinkan akses ke informasi sibuk bebas Anda. Pengguna ini juga dapat menjadwalkan acara tentatif langsung ke kalender Anda.',
+);