+++ www/modules/smime/language/id.php
@@ -1,0 +1,22 @@
+<?php
+return array (
+  'SMIME settings' => 'Pengaturan SMIME',
+  'PKCS12 certificate' => 'Sertifikat PKCS12',
+  'You have uploaded a certificate already. SMIME support is enabled for this account.' => 'Anda sudah mengunggah sertifikat. Dukungan SMIME diaktifkan untuk akun ini.',
+  'To upload a new PKCS12 certificate you must enter your Group-Office password. The Group-Office password must be different than your PCSK12 certificate for security reasons. No password is also prohibited.' => 'Untuk mengunggah sertifikat PKCS12 baru, Anda harus memasukkan kata sandi Group-Office. Kata sandi Group-Office harus berbeda dari sertifikat PCSK12 demi keamanan. Tidak ada kata sandi juga dilarang.',
+  'Select new PKCS12 Certificate' => 'Pilih Sertifikat PKCS12 baru',
+  'Delete certificate' => 'Hapus sertifikat',
+  'Download certificate' => 'Unduh sertifikat',
+  'Always sign messages' => 'Selalu tanda tangani pesan',
+  'SMIME Certificate' => 'Sertifikat SMIME',
+  'Public SMIME certificates' => 'Sertifikat SMIME publik',
+  'This message was sent to you encrypted.' => 'Pesan ini dikirimkan kepada Anda dengan dienkripsi.',
+  'This message is digitally signed. Click here to verify the signature and import the certificate.' => 'Pesan ini ditandatangani secara digital. Klik di sini untuk memverifikasi tanda tangan dan mengimpor sertifikat.',
+  'Add to SMIME certificates' => 'Tambah ke sertifikat SMIME',
+  'Certificate added' => 'Sertifikat ditambahkan',
+  'SMIME certificate was added for' => 'Sertifikat SMIME ditambahkan untuk',
+  'Could not add SMIME certificate for' => 'Tidak bisa menambahkan sertifikat SMIME untuk',
+  'Sign with SMIME' => 'Tanda tangani dengan SMIME',
+  'Encrypt with SMIME' => 'Enkripsi dengan SMIME',
+  'Please enter the password of your SMIME certificate.' => 'Masukkan kata sandi sertifikat SMIME Anda',
+);